# openapi/bundled.yaml
openapi: 3.1.0
info:
  title: Tally API (Bundled)
  version: 0.1.0
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
paths:
...
  /api/streaming-services:
    get:
      summary: List streaming services
      description: |
        Returns streaming services with price tiers for a given country.
        If `country` is omitted, the server falls back to the authenticated user's `users.country_code`, then 'US'.
      parameters:
        - name: country
          in: query
          required: false
          description: ISO 3166-1 alpha-2 country code (e.g., AU, US)
          schema:
            type: string
            minLength: 2
            maxLength: 2
            example: AU
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesResponse'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      services:
                        - id: c38c74f5-b020-4292-9aba-6342ba529485
                          tmdb_provider_id: 1899
                          name: HBO Max
                          logo_path: https://image.tmdb.org/t/p/w45/jbe4gVSfRlbPTdESXhEKpornsfu.jpg
                          homepage: https://www.max.com
                          prices:
                            - tier: Standard
                              amount: 14.99
                              currency: USD
                              billing_frequency: monthly
                              active: true
                              provider_name: HBO Max
                          default_price:
                            tier: Standard
                            amount: 14.99
                            currency: USD
                            billing_frequency: monthly
                            active: true
                      count: 1
        '401':
          $ref: '#/components/responses/responses-Unauthorized'
        '500':
          $ref: '#/components/responses/responses-ServerError'
  /api/streaming-services/{id}:
    get:
      summary: Get a streaming service by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      service:
                        id: c38c74f5-b020-4292-9aba-6342ba529485
                        name: HBO Max
                        logo_path: https://image.tmdb.org/t/p/w45/jbe4gVSfRlbPTdESXhEKpornsfu.jpg
                        homepage: https://www.max.com
        '401':
          $ref: '#/components/responses/responses-Unauthorized'
        '404':
          $ref: '#/components/responses/responses-NotFound'
        '500':
          $ref: '#/components/responses/responses-ServerError'
  /api/streaming-services/popular:
    get:
      summary: List most popular services by active subscriptions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopularServicesResponse'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      services:
                        - id: c38c74f5-b020-4292-9aba-6342ba529485
                          name: HBO Max
                          logo_path: https://image.tmdb.org/t/p/w45/jbe4gVSfRlbPTdESXhEKpornsfu.jpg
                          homepage: https://www.max.com
                          subscriber_count: 42
                      count: 1
        '401':
          $ref: '#/components/responses/responses-Unauthorized'
        '500':
          $ref: '#/components/responses/responses-ServerError'
  /api/streaming-services/regions:
    get:
      summary: List common regions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionsResponse'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      regions:
                        - code: US
                          name: United States
                        - code: AU
                          name: Australia
                        - code: GB
                          name: United Kingdom
                      count: 3
        '401':
          $ref: '#/components/responses/responses-Unauthorized'
        '500':
          $ref: '#/components/responses/responses-ServerError'
  /api/watchlist:
    get:
      summary: Get user's watchlist
      description: |
        Returns the user's watchlist with normalized show details and selected streaming provider.
        Optional `status` query filters by watchlist status.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - watchlist
              - watching
              - completed
              - dropped
              - all
            default: all
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistListResponse'
        '401':
          $ref: '#/components/responses/responses-Unauthorized'
        '500':
          $ref: '#/components/responses/responses-ServerError'
    post:
      summary: Add a show to the user's watchlist
      description: |
        Adds a show (by TMDB ID) to the user's watchlist; status defaults to `watchlist`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tmdbId
              properties:
                tmdbId:
                  type: integer
                status:
                  type: string
                  enum:
                    - watchlist
                    - watching
                  default: watchlist
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddWatchlistResponse'
        '400':
          $ref: '#/components/responses/responses-BadRequest'
        '401':
          $ref: '#/components/responses/responses-Unauthorized'
        '500':
          $ref: '#/components/responses/responses-ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: string
          nullable: true
...
    schemas-PriceTier:
      type: object
      properties:
        tier:
          type: string
        amount:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        billing_frequency:
          type: string
          nullable: true
          example: monthly
        active:
          type: boolean
          nullable: true
        notes:
          type: string
          nullable: true
        provider_name:
          type: string
          nullable: true
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tmdb_provider_id:
          type: integer
          nullable: true
        name:
          type: string
        logo_path:
          type: string
          nullable: true
        homepage:
          type: string
          nullable: true
        prices:
          type: array
          items:
            $ref: '#/components/schemas/schemas-PriceTier'
        default_price:
          $ref: '#/components/schemas/schemas-PriceTier'
    ServicesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            services:
              type: array
              items:
                $ref: '#/components/schemas/Service'
            count:
              type: integer
              example: 12
    ServiceResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            service:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                logo_path:
                  type: string
                  nullable: true
                homepage:
                  type: string
                  nullable: true
    PopularService:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logo_path:
          type: string
          nullable: true
        homepage:
          type: string
          nullable: true
        subscriber_count:
          type: integer
    PopularServicesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            services:
              type: array
              items:
                $ref: '#/components/schemas/PopularService'
            count:
              type: integer
              example: 10
    Region:
      type: object
      properties:
        code:
          type: string
          example: US
        name:
          type: string
          example: United States
    RegionsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            regions:
              type: array
              items:
                $ref: '#/components/schemas/Region'
            count:
              type: integer
              example: 10
    ShowSummary:
      type: object
      description: Normalized show object from DB + poster URL normalization
      properties:
        id:
          type: string
          format: uuid
        tmdb_id:
          type: integer
        title:
          type: string
        overview:
          type: string
          nullable: true
        poster_path:
          type: string
          nullable: true
        first_air_date:
          type: string
          format: date
          nullable: true
        last_air_date:
          type: string
          format: date
          nullable: true
        status:
          type: string
          nullable: true
        total_seasons:
          type: integer
          nullable: true
        total_episodes:
          type: integer
          nullable: true
    SelectedProvider:
      type: object
      required:
        - id
        - name
        - logo_path
      properties:
        id:
          type: integer
          description: TMDB provider id
        name:
          type: string
        logo_path:
          type: string
    WatchlistRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - watchlist
            - watching
            - completed
            - dropped
        show:
          $ref: '#/components/schemas/ShowSummary'
        streaming_provider:
          oneOf:
            - $ref: '#/components/schemas/SelectedProvider'
            - type: 'null'
    WatchlistListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            shows:
              type: array
              items:
                $ref: '#/components/schemas/WatchlistRow'
            totalCount:
              type: integer
            statusFilter:
              type: string
    UserShow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        show_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - watchlist
            - watching
            - completed
            - dropped
        added_at:
          type: string
          format: date-time
          nullable: true
        started_watching_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        last_episode_watched_id:
          type: string
          format: uuid
          nullable: true
        show_rating:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        buffer_days:
          type: integer
        selected_service_id:
          type: string
          format: uuid
          nullable: true
        country_code:
          type: string
          minLength: 2
          maxLength: 2
          nullable: true
    AddWatchlistResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            userShow:
              $ref: '#/components/schemas/UserShow'
            show:
              allOf:
                - $ref: '#/components/schemas/ShowSummary'
            availability:
              type: array
              items:
                type: object
                properties:
                  country_code:
                    type: string
                  service_id:
                    type: string
                    format: uuid
                  availability_type:
                    type: string
                  price_amount:
                    type: number
                    nullable: true
                  price_currency:
                    type: string
                    nullable: true
                  deep_link:
                    type: string
                    nullable: true
            message:
              type: string
...
  responses:
...
    responses-Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    responses-ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    responses-NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    responses-BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'